tweak code
